# integration_package/docs/guides/openapi-v72.yaml
openapi: 3.1.0
info:
  title: CHE·NU™ V72 API
  description: |
    Governed Intelligence Operating System API
    
    ## GOUVERNANCE > EXÉCUTION
    
    All sensitive actions require checkpoint approval.
    
    ### Principles
    - No autonomous agents
    - Human-in-the-loop for critical actions
    - Complete audit trail
    - READ-ONLY XR interface
  version: '72.0.0'
  contact:
    name: CHE·NU Support
    email: api@chenu.io
  license:
    name: Proprietary
    url: https://chenu.io/license
servers:
  - url: http://localhost:8000/api/v1
    description: Development
  - url: https://staging.chenu.io/api/v1
    description: Staging
  - url: https://api.chenu.io/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /spheres/{sphereId}:
    get:
      tags: [Spheres]
      summary: Get sphere by ID
      parameters:
        - $ref: '#/components/parameters/sphereId'
      responses:
        '200':
          description: Sphere details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SphereResponse'
...
  /agents:
    get:
      tags: [Agents]
      summary: List agents
      parameters:
        - name: level
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 3
        - name: domain
          in: query
          schema:
            type: string
        - name: hired_only
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentsResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    sphereId:
      name: sphereId
      in: path
      required: true
      schema:
        type: string
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Sphere:
      type: object
      properties:
        id:
          type: string
        name_fr:
          type: string
        name_en:
          type: string
        icon:
          type: string
        color:
          type: string
        order:
          type: integer
        stats:
          $ref: '#/components/schemas/SphereStats'
    SphereStats:
      type: object
      properties:
        threads_count:
          type: integer
        decisions_count:
          type: integer
        agents_count:
          type: integer
        files_count:
          type: integer
        meetings_count:
          type: integer
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name_fr:
          type: string
        name_en:
          type: string
        level:
          type: integer
          minimum: 0
          maximum: 3
        domain:
          type: string
        capabilities:
          type: array
          items:
            type: string
        cost:
          type: integer
        is_hired:
          type: boolean
...
    SphereResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Sphere'
...
    AgentsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
...